[REFACTOR]: Redis 테스트용 빈 패키지 위치 이동
